// packages/form/src/interface.ts
export type InternalNamePath = (string | number)[]
export type NamePath<T = any> = DeepNamePath<T>

export type StoreValue = any
export type Store = Record<string, StoreValue>

export interface Meta {
  touched: boolean
  validating: boolean
  errors: string[]
  warnings: string[]
  name: InternalNamePath
  validated: boolean
}

export interface InternalFieldData extends Meta {
  value: StoreValue
}

/**
 * Used by `setFields` config
 */
export interface FieldData<Values = any> extends Partial<Omit<InternalFieldData, 'name'>> {
  name: NamePath<Values>
}
// ...
export interface FieldError {
  name: InternalNamePath
  errors: string[]
  warnings: string[]
}

export interface RuleError {
  errors: string[]
  rule: RuleObject
}

export interface ValidateOptions {
  /**
   * Validate only and not trigger UI and Field status update
   */
  validateOnly?: boolean
  /**
   * Recursive validate. It will validate all the name path that contains the provided one.
   * e.g. [['a']] will validate ['a'] , ['a', 'b'] and ['a', 1].
   */
  recursive?: boolean
  /** Validate when a field is dirty (validated or touched) */
  dirty?: boolean
}

export interface ValidateFields<Values = any> {
  (opt?: ValidateOptions): Promise<Values>
  (nameList?: NamePath[], opt?: ValidateOptions): Promise<Values>
}
// ...
type RecursivePartial<T>
  = NonNullable<T> extends object
    ? {
        [P in keyof T]?: NonNullable<T[P]> extends (infer U)[]
          ? RecursivePartial<U>[]
          : NonNullable<T[P]> extends object
            ? RecursivePartial<T[P]>
            : T[P];
      }
    : T

export type FilterFunc = (meta: Meta | null) => boolean

export interface GetFieldsValueConfig {
  /**
   * @deprecated `strict` is deprecated and not working anymore
   */
  strict?: boolean
  filter?: FilterFunc
}

export interface FormInstance<Values = any> {
  // Origin Form API
  getFieldValue: (name: NamePath<Values>) => StoreValue
  getFieldsValue: (() => Values)
    & ((nameList: NamePath<Values>[] | true, filterFunc?: FilterFunc) => any)
    & ((config: GetFieldsValueConfig) => any)
  getFieldError: (name: NamePath<Values>) => string[]
  getFieldsError: (nameList?: NamePath<Values>[]) => FieldError[]
  getFieldWarning: (name: NamePath<Values>) => string[]
  isFieldsTouched: ((nameList?: NamePath<Values>[], allFieldsTouched?: boolean) => boolean)
    & ((allFieldsTouched?: boolean) => boolean)
  isFieldTouched: (name: NamePath<Values>) => boolean
  isFieldValidating: (name: NamePath<Values>) => boolean
  isFieldsValidating: (nameList?: NamePath<Values>[]) => boolean
  resetFields: (fields?: NamePath<Values>[]) => void
  setFields: (fields: FieldData<Values>[]) => void
  setFieldValue: (name: NamePath<Values>, value: any) => void
  setFieldsValue: (values: RecursivePartial<Values>) => void
  validateFields: ValidateFields<Values>

  // New API
  submit: () => void
}
// ...
type ValidateMessage = string | (() => string)
export interface ValidateMessages {
  default?: ValidateMessage
  required?: ValidateMessage
  enum?: ValidateMessage
  whitespace?: ValidateMessage
  date?: {
    format?: ValidateMessage
    parse?: ValidateMessage
    invalid?: ValidateMessage
  }
  types?: {
    string?: ValidateMessage
    method?: ValidateMessage
    array?: ValidateMessage
    object?: ValidateMessage
    number?: ValidateMessage
    date?: ValidateMessage
    boolean?: ValidateMessage
    integer?: ValidateMessage
    float?: ValidateMessage
    regexp?: ValidateMessage
    email?: ValidateMessage
    url?: ValidateMessage
    hex?: ValidateMessage
  }
  string?: {
    len?: ValidateMessage
    min?: ValidateMessage
    max?: ValidateMessage
    range?: ValidateMessage
  }
  number?: {
    len?: ValidateMessage
    min?: ValidateMessage
    max?: ValidateMessage
    range?: ValidateMessage
  }
  array?: {
    len?: ValidateMessage
    min?: ValidateMessage
    max?: ValidateMessage
    range?: ValidateMessage
  }
  pattern?: {
    mismatch?: ValidateMessage
  }
}